# statistics_manager_service/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: "This is the Statistics Manager Service OpenAPI definition. The Statistics\
    \ Manager Service processes energy data in order to provide the user with useful\
    \ statistics (e.g. energy cost for the day, injection for the month, consumption\
    \ for the year, CO2 footprint). The interactions here represented are based on\
    \ this [Canvas](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/Statistics-Manager-Service.adoc)"
  title: Statistics Manager Service
  version: 0.1.1
servers:
- url: https://interconnect-dev.inesctec.pt/api/statistics
security:
- {}
- Bearer: []
...
paths:
...
  /mock-consumption-curve-producer:
    post:
      description: Mocks a user's <user_id> consumption curve for a single day <start_date>.
        The curve consists of the energy consumption forecast and the real energy
        consumption. The curve's interval between values is given by <delivery_time>
        minutes.
      operationId: mock_consumption_curve_producer_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: form
      - explode: true
        in: query
        name: start_date
        required: true
        schema:
          $ref: '#/components/schemas/Timestamp'
        style: form
      - explode: true
        in: query
        name: delivery_time
        required: true
        schema:
          example: 60
          minimum: 15
          type: integer
        style: form
      responses:
        "200":
          description: Success response. Event has been triggered
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Trigger the service into saving a mock consumption curve (forecast
        + real) for a user
      tags:
      - Event Producers
      x-openapi-router-controller: statistics_manager_service.controllers.event_producers_controller
...
  /trigger-computed-forecast-producer:
    post:
      operationId: trigger_computed_forecast_producer_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: form
      - explode: true
        in: query
        name: forecast_day
        required: true
        schema:
          $ref: '#/components/schemas/Date'
        style: form
      responses:
        "200":
          description: Success response. Events have been triggered
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Trigger the service into saving a user's forecast data for a single
        day
      tags:
      - Event Producers
      x-openapi-router-controller: statistics_manager_service.controllers.event_producers_controller
  /users-get-real-data-producer:
    post:
      description: Produces a signal that triggers the service into saving a user's
        <user_id> real power consumption for a date interval. <start_date> and <end_date>
        are expected to be in UTC. The data retrieved is measured by the household
        smart meter + dongle (in case the household has one). The service retrieves
        the dongle data from influxdb and the smart meter data from the DSO interface.
      operationId: users_get_real_data_producer_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: true
        in: query
        name: user_id
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
        style: form
      - explode: true
        in: query
        name: start_datetime
        required: true
        schema:
          $ref: '#/components/schemas/Datetime'
        style: form
      - explode: true
        in: query
        name: end_datetime
        required: false
        schema:
          $ref: '#/components/schemas/Datetime'
        style: form
      responses:
        "200":
          description: Success response. Event has been triggered
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Timestamp error:
                  value:
                    error: Reason for bad request error
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "403":
          content:
            application/json:
              examples:
                Forbidden:
                  value:
                    error: you are not authorized to access this information
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden to access the information requested
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "404":
          content:
            application/json:
              examples:
                User Id not found:
                  value:
                    error: user with id "userId" not found
              schema:
                $ref: '#/components/schemas/Error'
          description: User id not found
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Triggers the service into saving a user's real power consumption for
        a date interval
      tags:
      - Event Producers
      x-openapi-router-controller: statistics_manager_service.controllers.event_producers_controller
...
components:
...
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
    UserId:
      example: userId1
      maxLength: 10
      minLength: 10
      title: user_id
      type: string
    Authorization:
      format: byte
      type: string
    Date:
      example: 2023-01-20
      format: date
      type: string
    Datetime:
      example: 2023-01-20 00:00:00
      format: datetime
      type: string
    Timestamp:
      example: 2023-01-20T00:00:00Z
      format: date-time
      title: timestamp
      type: string
...
    Error:
      description: Error model
      example:
        error: error
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: statistics_manager_service.controllers.security_controller.info_from_Bearer